AutoComplete: Rename class
