+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -6,7 +6,7 @@ import java.util.logging.Logger;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.ComponentManager;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.autocomplete.AutoCompleteManager;
+import seedu.address.logic.autocomplete.AutoComplete;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
@@ -28,15 +28,14 @@ public class LogicManager extends ComponentManager implements Logic {
     private final CommandHistory history;
     private final AddressBookParser addressBookParser;
     private final UndoRedoStack undoRedoStack;
-
-    private final AutoCompleteManager autoCompleteManager;
+    private final AutoComplete autoComplete;
 
     public LogicManager(Model model) {
         this.model = model;
         history = new CommandHistory();
         addressBookParser = new AddressBookParser();
         undoRedoStack = new UndoRedoStack();
-        autoCompleteManager = new AutoCompleteManager();
+        autoComplete = new AutoComplete();
     }
 
     @Override
@@ -80,6 +79,6 @@ public class LogicManager extends ComponentManager implements Logic {
 
     @Override
     public List<String> getListOfAutoComplete(String keyWord) {
-        return autoCompleteManager.getAutoCompleteWords(keyWord);
+        return autoComplete.getAutoCompleteWords(keyWord);
     }
 }
